+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # COFantasy: Historique des changements (depuis la version 1.0)
 ## 2.15
 ### Capacités
+* Charge fantastique du chevalier
+* Project de la Voie du colosse.
 * Ventre mou de la Voie du tueur de géants.
 * Suggestion de la Voie de la séduction du barde
 * Gestion du mot de pouvoir qui immobilise.
@@ -17,6 +19,7 @@
 * Injonction Mortelle (Magie Maléfique rang 3)
 
 ### Autres améliorations
+* Ajout d'une option de sélection alliesEnVue
 * Prise en compte des immunités pour les effets temporaires.
 * Attribut pour diviser par 2 les effets ou les dm d'un type donné
 * Possibilité d'avoir des bonus aux saves contre un type donné.
@@ -37,6 +40,7 @@
 * Ajout d'une option permettant l'affichage des coûts effectifs en PM au lancement du sort
 
 ### Corrections de bugs
+* Mise à jour de l'initiative quand on rengaine son arme à distance avec la voie du pistolero.
 * Donne le droit au joueur qui doit faire une réaction à une attaque de ne pas réagir.
 * Les créatures qui enveloppent ou étreignent leur cible la relachent quand elle meurt.
 * Le script ne prennait pas en compte la sagesse de l'attaquant contre les créatures immunisées aux armes.